+++ registry-ws/src/main/java/org/gbif/registry/ws/resources/DatasetResource.java
@@ -1013,7 +1013,6 @@ public class DatasetResource extends BaseNetworkEntityResource<Dataset>
   @Path("doi/{doi}")
   @Override
   public PagingResponse<Dataset> listByDOI(@PathParam("doi") String doi, @Context Pageable page) {
-    return new PagingResponse<>(page, Integer.valueOf(datasetMapper.countByDOI(doi)).longValue(),
-                                datasetMapper.listByDOI(doi, page));
+    return new PagingResponse<>(page, datasetMapper.countByDOI(doi), datasetMapper.listByDOI(doi, page));
   }
 }
